requirements.txt
pyautogui
Pillow
pytesseract
numpy